test(cli): split integration suite and extract helpers (M3)
Replace monolithic integration.test.ts with seven domain Vitest files.
Add test-support/integration-helpers.ts for engine factory, suite-scoped
temp dirs under .temp-integration/<slug>/, and shared FS helpers.

Addresses SLOBAC findings 4 (shared engine state) and 6 (monolithic file).
Memory bank: L2 build phase complete.
