--- packages/cli/test/integration/integration.test.ts
@@ -1,1507 +1,0 @@
-/**
- * Fixture-based integration tests for a16n.
- * 
- * Test fixtures follow the pattern:
- *   fixtures/<test-name>/from-<agent>/... (input files)
- *   fixtures/<test-name>/to-<agent>/... (expected output files)
- * 
- * The test copies from-* to a temp directory, runs conversion,
- * then compares the result with to-*.
- */
-import { describe, it, expect, beforeEach, afterEach } from 'vitest';
-import * as fs from 'fs/promises';
-import * as path from 'path';
-import { fileURLToPath } from 'url';
-import { A16nEngine } from '@a16njs/engine';
-import cursorPlugin from '@a16njs/plugin-cursor';
-import claudePlugin from '@a16njs/plugin-claude';
-import a16nPlugin from '@a16njs/plugin-a16n';
-import { WarningCode } from '@a16njs/models';
-
-const __dirname = path.dirname(fileURLToPath(import.meta.url));
-const fixturesDir = path.join(__dirname, 'fixtures');
-const tempDir = path.join(__dirname, '.temp-integration');
-
-// Create engine with all three plugins
-const engine = new A16nEngine([cursorPlugin, claudePlugin, a16nPlugin]);
-
-/**
- * Recursively copy a directory.
- */
-async function copyDir(src: string, dest: string): Promise<void> {
-  await fs.mkdir(dest, { recursive: true });
-  const entries = await fs.readdir(src, { withFileTypes: true });
-  
-  for (const entry of entries) {
-    const srcPath = path.join(src, entry.name);
-    const destPath = path.join(dest, entry.name);
-    
-    if (entry.isDirectory()) {
-      await copyDir(srcPath, destPath);
-    } else {
-      await fs.copyFile(srcPath, destPath);
-    }
-  }
-}
-
-/**
- * Recursively read all files in a directory into a map.
- */
-async function readDirFiles(dir: string, base: string = ''): Promise<Map<string, string>> {
-  const files = new Map<string, string>();
-  
-  try {
-    const entries = await fs.readdir(dir, { withFileTypes: true });
-    
-    for (const entry of entries) {
-      const relativePath = base ? `${base}/${entry.name}` : entry.name;
-      const fullPath = path.join(dir, entry.name);
-      
-      if (entry.isDirectory()) {
-        const subFiles = await readDirFiles(fullPath, relativePath);
-        for (const [k, v] of subFiles) {
-          files.set(k, v);
-        }
-      } else {
-        const content = await fs.readFile(fullPath, 'utf-8');
-        files.set(relativePath, content);
-      }
-    }
-  } catch {
-    // Directory doesn't exist
-  }
-  
-  return files;
-}
-
-/**
- * Compare actual output with expected output.
- */
-function compareOutputs(actual: Map<string, string>, expected: Map<string, string>): void {
-  // Check all expected files exist with correct content
-  for (const [path, expectedContent] of expected) {
-    const actualContent = actual.get(path);
-    expect(actualContent, `Expected file ${path} to exist`).toBeDefined();
-    expect(actualContent?.trim(), `Content mismatch in ${path}`).toBe(expectedContent.trim());
-  }
-}
-
-describe('Integration Tests - Fixture Based', () => {
-  beforeEach(async () => {
-    await fs.rm(tempDir, { recursive: true, force: true });
-    await fs.mkdir(tempDir, { recursive: true });
-  });
-
-  afterEach(async () => {
-    await fs.rm(tempDir, { recursive: true, force: true });
-  });
-
-  describe('cursor-to-claude-basic', () => {
-    it('should convert a single Cursor rule to Claude format', async () => {
-      const fixturePath = path.join(fixturesDir, 'cursor-to-claude-basic');
-      const fromDir = path.join(fixturePath, 'from-cursor');
-      const toDir = path.join(fixturePath, 'to-claude');
-      
-      // Copy input to temp
-      await copyDir(fromDir, tempDir);
-      
-      // Run conversion
-      const result = await engine.convert({
-        source: 'cursor',
-        target: 'claude',
-        root: tempDir,
-      });
-      
-      // Verify conversion succeeded
-      expect(result.discovered.length).toBeGreaterThan(0);
-      expect(result.written.length).toBeGreaterThan(0);
-      
-      // Read actual and expected outputs
-      const actualRulesDir = path.join(tempDir, '.claude', 'rules');
-      const actualFiles = await readDirFiles(actualRulesDir);
-      
-      // Verify rule file was created
-      expect(actualFiles.size).toBeGreaterThan(0);
-      
-      // Check content includes expected text from cursor rule
-      const ruleContent = Array.from(actualFiles.values())[0];
-      expect(ruleContent).toContain('Always use TypeScript');
-    });
-  });
-
-  describe('claude-to-cursor-basic', () => {
-    it('should convert CLAUDE.md to a Cursor rule', async () => {
-      const fixturePath = path.join(fixturesDir, 'claude-to-cursor-basic');
-      const fromDir = path.join(fixturePath, 'from-claude');
-      const toDir = path.join(fixturePath, 'to-cursor');
-      
-      // Copy input to temp
-      await copyDir(fromDir, tempDir);
-      
-      // Run conversion
-      const result = await engine.convert({
-        source: 'claude',
-        target: 'cursor',
-        root: tempDir,
-      });
-      
-      // Verify conversion succeeded
-      expect(result.discovered.length).toBeGreaterThan(0);
-      expect(result.written.length).toBeGreaterThan(0);
-      
-      // Read actual outputs
-      const rulesDir = path.join(tempDir, '.cursor', 'rules');
-      const actualFiles = await readDirFiles(rulesDir);
-      
-      // Verify a rule file was created
-      expect(actualFiles.size).toBeGreaterThan(0);
-      
-      // Check the content includes the original text
-      const ruleContent = Array.from(actualFiles.values())[0];
-      expect(ruleContent).toContain('alwaysApply: true');
-      expect(ruleContent).toContain('Write tests for all functions');
-      expect(ruleContent).toContain('Follow DRY principles');
-    });
-  });
-
-  describe('cursor-to-claude-multiple', () => {
-    it('should emit multiple Cursor rules as separate .claude/rules/*.md files', async () => {
-      const fixturePath = path.join(fixturesDir, 'cursor-to-claude-multiple');
-      const fromDir = path.join(fixturePath, 'from-cursor');
-      
-      // Copy input to temp
-      await copyDir(fromDir, tempDir);
-      
-      // Run conversion
-      const result = await engine.convert({
-        source: 'cursor',
-        target: 'claude',
-        root: tempDir,
-      });
-      
-      // Verify multiple items discovered
-      expect(result.discovered.length).toBe(2);
-      // BREAKING: Each gets its own file now
-      expect(result.written.length).toBe(2);
-      
-      // BREAKING: Should NOT have a Merged warning
-      expect(result.warnings.some(w => w.code === 'merged')).toBe(false);
-      
-      // Read actual rule files
-      const rulesDir = path.join(tempDir, '.claude', 'rules');
-      const files = await fs.readdir(rulesDir);
-      expect(files).toHaveLength(2);
-      
-      // Read both files and verify both rules are present
-      const contents = await Promise.all(
-        files.map(f => fs.readFile(path.join(rulesDir, f), 'utf-8'))
-      );
-      const combinedContent = contents.join('\n');
-      
-      // Should contain both rules (in separate files)
-      expect(combinedContent).toContain('Use 2-space indentation');
-      expect(combinedContent).toContain('Write unit tests');
-    });
-  });
-
-  describe('dry-run mode', () => {
-    it('should not write files in dry-run mode', async () => {
-      const fixturePath = path.join(fixturesDir, 'cursor-to-claude-basic');
-      const fromDir = path.join(fixturePath, 'from-cursor');
-      
-      // Copy input to temp
-      await copyDir(fromDir, tempDir);
-      
-      // Run conversion in dry-run mode
-      const result = await engine.convert({
-        source: 'cursor',
-        target: 'claude',
-        root: tempDir,
-        dryRun: true,
-      });
-      
-      // Should discover items and show what would be written (without actually writing)
-      expect(result.discovered.length).toBeGreaterThan(0);
-      expect(result.written.length).toBeGreaterThan(0); // Now returns what WOULD be written
-      
-      // But .claude/rules/ should NOT actually exist (files not written in dry-run)
-      await expect(
-        fs.access(path.join(tempDir, '.claude', 'rules'))
-      ).rejects.toThrow();
-    });
-  });
-
-  describe('empty project', () => {
-    it('should handle project with no rules', async () => {
-      // Empty temp dir (no rules)
-      const result = await engine.convert({
-        source: 'cursor',
-        target: 'claude',
-        root: tempDir,
-      });
-      
-      expect(result.discovered.length).toBe(0);
-      expect(result.written.length).toBe(0);
-      expect(result.warnings.length).toBe(0);
-    });
-  });
-});
-
-describe('Integration Tests - FileRule and SimpleAgentSkill', () => {
-  beforeEach(async () => {
-    await fs.rm(tempDir, { recursive: true, force: true });
-    await fs.mkdir(tempDir, { recursive: true });
-  });
-
-  afterEach(async () => {
-    await fs.rm(tempDir, { recursive: true, force: true });
-  });
-
-  describe('cursor-filerule-to-claude', () => {
-    it('should convert Cursor FileRule to Claude native rule file', async () => {
-      const fixturePath = path.join(fixturesDir, 'cursor-filerule-to-claude');
-      const fromDir = path.join(fixturePath, 'from-cursor');
-      
-      // Copy input to temp
-      await copyDir(fromDir, tempDir);
-      
-      // Run conversion
-      const result = await engine.convert({
-        source: 'cursor',
-        target: 'claude',
-        root: tempDir,
-      });
-      
-      // Verify FileRule was discovered
-      const fileRules = result.discovered.filter(d => d.type === 'file-rule');
-      expect(fileRules).toHaveLength(1);
-      
-      // Verify rule file was created in .claude/rules/
-      const ruleContent = await fs.readFile(
-        path.join(tempDir, '.claude', 'rules', 'react.md'),
-        'utf-8'
-      );
-      expect(ruleContent).toContain('Use React best practices');
-      expect(ruleContent).toContain('paths:');
-      expect(ruleContent).toContain('**/*.tsx');
-      
-      // BREAKING: settings.local.json should NOT exist (native rules now)
-      await expect(
-        fs.access(path.join(tempDir, '.claude', 'settings.local.json'))
-      ).rejects.toThrow();
-      
-      // BREAKING: NO approximation warning (native support now)
-      const approxWarning = result.warnings.find(w => w.code === 'approximated');
-      expect(approxWarning).toBeUndefined();
-    });
-  });
-
-  describe('cursor-agentskill-to-claude', () => {
-    it('should convert Cursor SimpleAgentSkill to Claude skill', async () => {
-      const fixturePath = path.join(fixturesDir, 'cursor-agentskill-to-claude');
-      const fromDir = path.join(fixturePath, 'from-cursor');
-      
-      // Copy input to temp
-      await copyDir(fromDir, tempDir);
-      
-      // Run conversion
-      const result = await engine.convert({
-        source: 'cursor',
-        target: 'claude',
-        root: tempDir,
-      });
-      
-      // Verify SimpleAgentSkill was discovered
-      const agentSkills = result.discovered.filter(d => d.type === 'simple-agent-skill');
-      expect(agentSkills).toHaveLength(1);
-      
-      // Verify skill file was created
-      const skillContent = await fs.readFile(
-        path.join(tempDir, '.claude', 'skills', 'auth', 'SKILL.md'),
-        'utf-8'
-      );
-      // Description is quoted for YAML safety
-      expect(skillContent).toContain('description: "Authentication and authorization helper"');
-      expect(skillContent).toContain('Use JWT for stateless authentication');
-    });
-  });
-
-  describe('claude-skill-to-cursor', () => {
-    it('should convert Claude skill to Cursor rule with description', async () => {
-      const fixturePath = path.join(fixturesDir, 'claude-skill-to-cursor');
-      const fromDir = path.join(fixturePath, 'from-claude');
-      
-      // Copy input to temp
-      await copyDir(fromDir, tempDir);
-      
-      // Run conversion
-      const result = await engine.convert({
-        source: 'claude',
-        target: 'cursor',
-        root: tempDir,
-      });
-      
-      // Verify SimpleAgentSkill was discovered
-      const agentSkills = result.discovered.filter(d => d.type === 'simple-agent-skill');
-      expect(agentSkills).toHaveLength(1);
-      
-      // Read the output skill files (Phase 7: SimpleAgentSkill → .cursor/skills/)
-      const skillsDir = path.join(tempDir, '.cursor', 'skills');
-      const dirs = await fs.readdir(skillsDir);
-      expect(dirs).toHaveLength(1);
-      
-      const skillContent = await fs.readFile(
-        path.join(skillsDir, dirs[0]!, 'SKILL.md'),
-        'utf-8'
-      );
-      expect(skillContent).toContain('description:');
-      expect(skillContent).toContain('Testing best practices');
-      expect(skillContent).toContain('Write unit tests first');
-    });
-  });
-
-  describe('cursor-skill-names-to-claude', () => {
-    it('should preserve skill directory names (invocation names) through conversion', async () => {
-      const fixturePath = path.join(fixturesDir, 'cursor-skill-names-to-claude');
-      const fromDir = path.join(fixturePath, 'from-cursor');
-
-      await copyDir(fromDir, tempDir);
-
-      const result = await engine.convert({
-        source: 'cursor',
-        target: 'claude',
-        root: tempDir,
-      });
-
-      const agentSkills = result.discovered.filter(d => d.type === 'simple-agent-skill');
-      expect(agentSkills).toHaveLength(2);
-
-      // Skill directory names should match their original invocation names
-      const bananaSkill = await fs.readFile(
-        path.join(tempDir, '.claude', 'skills', 'banana', 'SKILL.md'),
-        'utf-8'
-      );
-      expect(bananaSkill).toContain('Print a banana emoji.');
-
-      const tomatoSkill = await fs.readFile(
-        path.join(tempDir, '.claude', 'skills', 'tomato', 'SKILL.md'),
-        'utf-8'
-      );
-      expect(tomatoSkill).toContain('Print a tomato emoji.');
-    });
-
-    it('should discover nested skills under category directories', async () => {
-      const fixturePath = path.join(fixturesDir, 'cursor-skill-names-to-claude');
-      const fromDir = path.join(fixturePath, 'from-cursor');
-
-      await copyDir(fromDir, tempDir);
-
-      const result = await engine.convert({
-        source: 'cursor',
-        target: 'claude',
-        root: tempDir,
-      });
-
-      // Both banana (top-level) and tomato (nested under veggies/) should be discovered
-      const agentSkills = result.discovered.filter(d => d.type === 'simple-agent-skill');
-      expect(agentSkills).toHaveLength(2);
-
-      // Verify nested skill identity — tomato originates from veggies/ subdirectory
-      const tomatoSkill = agentSkills.find(d => d.sourcePath?.includes('veggies/'));
-      expect(tomatoSkill).toBeDefined();
-      expect(tomatoSkill?.sourcePath).toContain('tomato');
-    });
-  });
-});
-
-describe('Integration Tests - AgentIgnore', () => {
-  beforeEach(async () => {
-    await fs.rm(tempDir, { recursive: true, force: true });
-    await fs.mkdir(tempDir, { recursive: true });
-  });
-
-  afterEach(async () => {
-    await fs.rm(tempDir, { recursive: true, force: true });
-  });
-
-  describe('cursor-ignore-to-claude', () => {
-    it('should convert .cursorignore to permissions.deny in settings.json', async () => {
-      const fixturePath = path.join(fixturesDir, 'cursor-ignore-to-claude');
-      const fromDir = path.join(fixturePath, 'from-cursor');
-      
-      // Copy input to temp
-      await copyDir(fromDir, tempDir);
-      
-      // Run conversion
-      const result = await engine.convert({
-        source: 'cursor',
-        target: 'claude',
-        root: tempDir,
-      });
-      
-      // Verify AgentIgnore was discovered
-      const agentIgnores = result.discovered.filter(d => d.type === 'agent-ignore');
-      expect(agentIgnores).toHaveLength(1);
-      
-      // Verify settings.json was created with permissions.deny
-      const settings = JSON.parse(
-        await fs.readFile(
-          path.join(tempDir, '.claude', 'settings.json'),
-          'utf-8'
-        )
-      );
-      expect(settings.permissions).toBeDefined();
-      expect(settings.permissions.deny).toBeInstanceOf(Array);
-      expect(settings.permissions.deny).toContain('Read(./dist/**)');
-      expect(settings.permissions.deny).toContain('Read(./.env)');
-      expect(settings.permissions.deny).toContain('Read(./**/*.log)');
-      
-      // Verify approximation warning was emitted
-      const approxWarning = result.warnings.find(w => w.code === 'approximated');
-      expect(approxWarning).toBeDefined();
-      expect(approxWarning?.message).toContain('permissions.deny');
-    });
-  });
-
-  describe('claude-ignore-to-cursor', () => {
-    it('should convert permissions.deny Read rules to .cursorignore', async () => {
-      const fixturePath = path.join(fixturesDir, 'claude-ignore-to-cursor');
-      const fromDir = path.join(fixturePath, 'from-claude');
-      
-      // Copy input to temp
-      await copyDir(fromDir, tempDir);
-      
-      // Run conversion
-      const result = await engine.convert({
-        source: 'claude',
-        target: 'cursor',
-        root: tempDir,
-      });
-      
-      // Verify AgentIgnore was discovered
-      const agentIgnores = result.discovered.filter(d => d.type === 'agent-ignore');
-      expect(agentIgnores).toHaveLength(1);
-      
-      // Verify .cursorignore was created
-      const ignoreContent = await fs.readFile(
-        path.join(tempDir, '.cursorignore'),
-        'utf-8'
-      );
-      expect(ignoreContent).toContain('dist/');
-      expect(ignoreContent).toContain('.env');
-      expect(ignoreContent).toContain('*.log');
-    });
-  });
-
-  describe('round-trip: cursor -> claude -> cursor', () => {
-    it('should preserve patterns through round-trip conversion', async () => {
-      const fixturePath = path.join(fixturesDir, 'cursor-ignore-to-claude');
-      const fromDir = path.join(fixturePath, 'from-cursor');
-      
-      // Copy input to temp
-      await copyDir(fromDir, tempDir);
-      
-      // Save original .cursorignore
-      const originalIgnore = await fs.readFile(
-        path.join(tempDir, '.cursorignore'),
-        'utf-8'
-      );
-      const originalPatterns = originalIgnore
-        .split('\n')
-        .map(l => l.trim())
-        .filter(l => l && !l.startsWith('#'));
-      
-      // Convert Cursor -> Claude
-      await engine.convert({
-        source: 'cursor',
-        target: 'claude',
-        root: tempDir,
-      });
-      
-      // Delete original .cursorignore to avoid confusion
-      await fs.unlink(path.join(tempDir, '.cursorignore'));
-      
-      // Convert Claude -> Cursor
-      await engine.convert({
-        source: 'claude',
-        target: 'cursor',
-        root: tempDir,
-      });
-      
-      // Read new .cursorignore
-      const newIgnore = await fs.readFile(
-        path.join(tempDir, '.cursorignore'),
-        'utf-8'
-      );
-      const newPatterns = newIgnore
-        .split('\n')
-        .map(l => l.trim())
-        .filter(l => l && !l.startsWith('#'));
-      
-      // Verify all original patterns are preserved
-      for (const pattern of originalPatterns) {
-        expect(newPatterns).toContain(pattern);
-      }
-    });
-  });
-});
-
-describe('Integration Tests - ManualPrompt (Commands)', () => {
-  beforeEach(async () => {
-    await fs.rm(tempDir, { recursive: true, force: true });
-    await fs.mkdir(tempDir, { recursive: true });
-  });
-
-  afterEach(async () => {
-    await fs.rm(tempDir, { recursive: true, force: true });
-  });
-
-  describe('cursor-command-to-claude', () => {
-    it('should convert simple Cursor command to Claude skill', async () => {
-      const fixturePath = path.join(fixturesDir, 'cursor-command-to-claude');
-      const fromDir = path.join(fixturePath, 'from-cursor');
-      
-      // Copy input to temp
-      await copyDir(fromDir, tempDir);
-      
-      // Run conversion
-      const result = await engine.convert({
-        source: 'cursor',
-        target: 'claude',
-        root: tempDir,
-      });
-      
-      // Verify ManualPrompt was discovered
-      const manualPrompts = result.discovered.filter(d => d.type === 'manual-prompt');
-      expect(manualPrompts).toHaveLength(1);
-      
-      // Verify skill file was created
-      const skillContent = await fs.readFile(
-        path.join(tempDir, '.claude', 'skills', 'review', 'SKILL.md'),
-        'utf-8'
-      );
-      expect(skillContent).toContain('name: "review"');
-      expect(skillContent).toContain('description: "Invoke with /review"');
-      expect(skillContent).toContain('Security vulnerabilities');
-    });
-  });
-
-  describe('cursor-command-complex-skipped', () => {
-    it('should skip complex commands and emit warning', async () => {
-      // Create a fixture with complex commands
-      await fs.mkdir(path.join(tempDir, '.cursor', 'commands'), { recursive: true });
-      
-      // Complex command with $ARGUMENTS
-      await fs.writeFile(
-        path.join(tempDir, '.cursor', 'commands', 'fix-issue.md'),
-        'Fix issue #$ARGUMENTS following best practices.',
-        'utf-8'
-      );
-      
-      // Run conversion
-      const result = await engine.convert({
-        source: 'cursor',
-        target: 'claude',
-        root: tempDir,
-      });
-      
-      // No commands should be discovered (complex is skipped)
-      const manualPrompts = result.discovered.filter(d => d.type === 'manual-prompt');
-      expect(manualPrompts).toHaveLength(0);
-      
-      // Should have a skipped warning
-      const skippedWarning = result.warnings.find(
-        w => w.code === 'skipped' && w.message.includes('fix-issue')
-      );
-      expect(skippedWarning).toBeDefined();
-      expect(skippedWarning?.message).toContain('$ARGUMENTS');
-    });
-  });
-
-  describe('cursor-to-cursor-command-passthrough', () => {
-    it('should preserve commands in cursor-to-cursor conversion', async () => {
-      // Create input commands
-      await fs.mkdir(path.join(tempDir, '.cursor', 'commands'), { recursive: true });
-      
-      const commandContent = 'Review this code for security.';
-      await fs.writeFile(
-        path.join(tempDir, '.cursor', 'commands', 'review.md'),
-        commandContent,
-        'utf-8'
-      );
-      
-      // Run cursor-to-cursor conversion (effectively a round-trip)
-      const result = await engine.convert({
-        source: 'cursor',
-        target: 'cursor',
-        root: tempDir,
-      });
-      
-      // Command should be discovered and written back as ManualPrompt
-      const manualPrompts = result.discovered.filter(d => d.type === 'manual-prompt');
-      expect(manualPrompts).toHaveLength(1);
-      
-      // Verify file was written
-      const outputContent = await fs.readFile(
-        path.join(tempDir, '.cursor', 'commands', 'review.md'),
-        'utf-8'
-      );
-      expect(outputContent).toBe(commandContent);
-    });
-  });
-
-  describe('claude-to-cursor-no-commands', () => {
-    it('should not produce ManualPrompts when converting from Claude (no skills with disable-model-invocation)', async () => {
-      // Create Claude input with skills (but no commands - Claude has no command concept)
-      await fs.mkdir(path.join(tempDir, '.claude', 'skills', 'testing'), { recursive: true });
-      await fs.writeFile(
-        path.join(tempDir, '.claude', 'skills', 'testing', 'SKILL.md'),
-        `---
-description: "Testing best practices"
----
-
-Write unit tests first.
-`,
-        'utf-8'
-      );
-      
-      // Run conversion
-      const result = await engine.convert({
-        source: 'claude',
-        target: 'cursor',
-        root: tempDir,
-      });
-      
-      // No ManualPrompts should be discovered (Claude has no command concept)
-      const manualPrompts = result.discovered.filter(d => d.type === 'manual-prompt');
-      expect(manualPrompts).toHaveLength(0);
-      
-      // .cursor/commands directory should NOT exist
-      await expect(
-        fs.access(path.join(tempDir, '.cursor', 'commands'))
-      ).rejects.toThrow();
-    });
-  });
-
-  describe('cursor-to-claude-complex-skill', () => {
-    it('should convert complex Cursor skill with resources to Claude', async () => {
-      const fixturePath = path.join(fixturesDir, 'cursor-to-claude-complex-skill');
-      const fromDir = path.join(fixturePath, 'from-cursor');
-      const toDir = path.join(fixturePath, 'to-claude');
-      
-      // Copy input to temp
-      await copyDir(fromDir, tempDir);
-      
-      // Run conversion
-      const result = await engine.convert({
-        source: 'cursor',
-        target: 'claude',
-        root: tempDir,
-      });
-      
-      // Verify conversion succeeded
-      expect(result.discovered.length).toBe(1);
-      expect(result.discovered[0].type).toBe('agent-skill-io');
-      expect(result.written.length).toBeGreaterThan(0);
-      
-      // Read actual and expected outputs
-      const actualSkillDir = path.join(tempDir, '.claude', 'skills', 'deploy');
-      const expectedSkillDir = path.join(toDir, '.claude', 'skills', 'deploy');
-      
-      const actualFiles = await readDirFiles(actualSkillDir);
-      const expectedFiles = await readDirFiles(expectedSkillDir);
-      
-      // Verify all files were created
-      compareOutputs(actualFiles, expectedFiles);
-      
-      // Verify resource files exist
-      expect(actualFiles.has('SKILL.md')).toBe(true);
-      expect(actualFiles.has('checklist.md')).toBe(true);
-      expect(actualFiles.has('config.json')).toBe(true);
-    });
-  });
-
-  describe('roundtrip-cursor-complex', () => {
-    it('should preserve complex Cursor skill through round-trip conversion', async () => {
-      const fixturePath = path.join(fixturesDir, 'roundtrip-cursor-complex');
-      const fromDir = path.join(fixturePath, 'from-cursor');
-      const toDir = path.join(fixturePath, 'to-cursor');
-      
-      // Copy input to temp
-      await copyDir(fromDir, tempDir);
-      
-      // Convert Cursor → Claude
-      const result1 = await engine.convert({
-        source: 'cursor',
-        target: 'claude',
-        root: tempDir,
-      });
-      
-      expect(result1.discovered.length).toBe(1);
-      expect(result1.discovered[0].type).toBe('agent-skill-io');
-      
-      // Convert Claude → Cursor (back)
-      const result2 = await engine.convert({
-        source: 'claude',
-        target: 'cursor',
-        root: tempDir,
-      });
-      
-      expect(result2.discovered.length).toBe(1);
-      expect(result2.discovered[0].type).toBe('agent-skill-io');
-      
-      // Read actual and expected outputs
-      const actualSkillDir = path.join(tempDir, '.cursor', 'skills', 'database');
-      const expectedSkillDir = path.join(toDir, '.cursor', 'skills', 'database');
-      
-      const actualFiles = await readDirFiles(actualSkillDir);
-      const expectedFiles = await readDirFiles(expectedSkillDir);
-      
-      // Verify all files preserved through round-trip
-      compareOutputs(actualFiles, expectedFiles);
-      
-      // Verify all resource files exist
-      expect(actualFiles.has('SKILL.md')).toBe(true);
-      expect(actualFiles.has('schema.sql')).toBe(true);
-      expect(actualFiles.has('migrations.md')).toBe(true);
-    });
-  });
-
-});
-
-describe('Integration Tests - Split Directories (--from-dir / --to-dir)', () => {
-  beforeEach(async () => {
-    await fs.rm(tempDir, { recursive: true, force: true });
-    await fs.mkdir(tempDir, { recursive: true });
-  });
-
-  afterEach(async () => {
-    await fs.rm(tempDir, { recursive: true, force: true });
-  });
-
-  it('convert with sourceRoot reads from specified source, writes to default root', async () => {
-    const sourceDir = path.join(tempDir, 'source');
-    await fs.mkdir(path.join(sourceDir, '.cursor', 'rules'), { recursive: true });
-    await fs.writeFile(
-      path.join(sourceDir, '.cursor/rules/coding.mdc'),
-      '---\nalwaysApply: true\n---\n\nAlways use TypeScript.'
-    );
-
-    const result = await engine.convert({
-      source: 'cursor',
-      target: 'claude',
-      root: tempDir,
-      sourceRoot: sourceDir,
-    });
-
-    expect(result.discovered).toHaveLength(1);
-    expect(result.written).toHaveLength(1);
-    // Output should be in tempDir (the default root), not sourceDir
-    expect(result.written[0]!.path).toContain(tempDir);
-    expect(result.written[0]!.path).not.toContain(path.join(tempDir, 'source'));
-    // Verify file was created
-    const rulesDir = path.join(tempDir, '.claude', 'rules');
-    const files = await fs.readdir(rulesDir);
-    expect(files.length).toBeGreaterThan(0);
-  });
-
-  it('convert with targetRoot reads from default root, writes to specified target', async () => {
-    const targetDir = path.join(tempDir, 'target');
-    await fs.mkdir(targetDir, { recursive: true });
-    await fs.mkdir(path.join(tempDir, '.cursor', 'rules'), { recursive: true });
-    await fs.writeFile(
-      path.join(tempDir, '.cursor/rules/coding.mdc'),
-      '---\nalwaysApply: true\n---\n\nAlways use TypeScript.'
-    );
-
-    const result = await engine.convert({
-      source: 'cursor',
-      target: 'claude',
-      root: tempDir,
-      targetRoot: targetDir,
-    });
-
-    expect(result.discovered).toHaveLength(1);
-    expect(result.written).toHaveLength(1);
-    // Output should be in targetDir
-    expect(result.written[0]!.path).toContain(targetDir);
-    // Verify file content
-    const rulesDir = path.join(targetDir, '.claude', 'rules');
-    const files = await fs.readdir(rulesDir);
-    expect(files.length).toBeGreaterThan(0);
-    const content = await fs.readFile(path.join(rulesDir, files[0]!), 'utf-8');
-    expect(content).toContain('Always use TypeScript');
-  });
-
-  it('convert with both sourceRoot and targetRoot', async () => {
-    const sourceDir = path.join(tempDir, 'src');
-    const targetDir = path.join(tempDir, 'out');
-    await fs.mkdir(path.join(sourceDir, '.cursor', 'rules'), { recursive: true });
-    await fs.mkdir(targetDir, { recursive: true });
-    await fs.writeFile(
-      path.join(sourceDir, '.cursor/rules/style.mdc'),
-      '---\nalwaysApply: true\n---\n\nUse 2-space indentation.'
-    );
-
-    const result = await engine.convert({
-      source: 'cursor',
-      target: 'claude',
-      root: '/unused',
-      sourceRoot: sourceDir,
-      targetRoot: targetDir,
-    });
-
-    expect(result.discovered).toHaveLength(1);
-    expect(result.written).toHaveLength(1);
-    expect(result.written[0]!.path).toContain(targetDir);
-    // tempDir root should NOT have output
-    await expect(fs.access(path.join(tempDir, '.claude'))).rejects.toThrow();
-    // Verify content
-    const rulesDir = path.join(targetDir, '.claude', 'rules');
-    const content = await fs.readFile(path.join(rulesDir, 'style.md'), 'utf-8');
-    expect(content).toContain('Use 2-space indentation');
-  });
-});
-
-describe('Integration Tests - Path Reference Rewriting (--rewrite-path-refs)', () => {
-  beforeEach(async () => {
-    await fs.rm(tempDir, { recursive: true, force: true });
-    await fs.mkdir(tempDir, { recursive: true });
-  });
-
-  afterEach(async () => {
-    await fs.rm(tempDir, { recursive: true, force: true });
-  });
-
-  it('Cursor→Claude with rewritePathRefs rewrites .cursor/rules/... → .claude/rules/...', async () => {
-    // Create cursor rules that reference each other
-    await fs.mkdir(path.join(tempDir, '.cursor', 'rules'), { recursive: true });
-    await fs.writeFile(
-      path.join(tempDir, '.cursor/rules/main.mdc'),
-      '---\nalwaysApply: true\n---\n\nLoad: .cursor/rules/auth.mdc\nLoad: .cursor/rules/db.mdc'
-    );
-    await fs.writeFile(
-      path.join(tempDir, '.cursor/rules/auth.mdc'),
-      '---\nalwaysApply: true\n---\n\nUse JWT for authentication.\nSee also: .cursor/rules/main.mdc'
-    );
-    await fs.writeFile(
-      path.join(tempDir, '.cursor/rules/db.mdc'),
-      '---\nalwaysApply: true\n---\n\nUse PostgreSQL.'
-    );
-
-    const result = await engine.convert({
-      source: 'cursor',
-      target: 'claude',
-      root: tempDir,
-      rewritePathRefs: true,
-    });
-
-    expect(result.discovered).toHaveLength(3);
-    expect(result.written).toHaveLength(3);
-
-    // Read main.md — should have rewritten references
-    const mainContent = await fs.readFile(
-      path.join(tempDir, '.claude/rules/main.md'),
-      'utf-8'
-    );
-    expect(mainContent).toContain('.claude/rules/auth.md');
-    expect(mainContent).toContain('.claude/rules/db.md');
-    expect(mainContent).not.toContain('.cursor/rules/auth.mdc');
-    expect(mainContent).not.toContain('.cursor/rules/db.mdc');
-
-    // Read auth.md — should have back-reference rewritten
-    const authContent = await fs.readFile(
-      path.join(tempDir, '.claude/rules/auth.md'),
-      'utf-8'
-    );
-    expect(authContent).toContain('.claude/rules/main.md');
-    expect(authContent).not.toContain('.cursor/rules/main.mdc');
-  });
-
-  it('Cursor→Claude with rewritePathRefs warns about orphan refs', async () => {
-    // Create a cursor rule that references a nonexistent rule
-    await fs.mkdir(path.join(tempDir, '.cursor', 'rules'), { recursive: true });
-    await fs.writeFile(
-      path.join(tempDir, '.cursor/rules/main.mdc'),
-      '---\nalwaysApply: true\n---\n\nSee .cursor/rules/missing-rule.mdc'
-    );
-
-    const result = await engine.convert({
-      source: 'cursor',
-      target: 'claude',
-      root: tempDir,
-      rewritePathRefs: true,
-    });
-
-    // Should have an OrphanPathRef warning
-    const orphanWarnings = result.warnings.filter(w => w.code === WarningCode.OrphanPathRef);
-    expect(orphanWarnings.length).toBeGreaterThanOrEqual(1);
-    expect(orphanWarnings[0]!.message).toContain('.cursor/rules/missing-rule.mdc');
-  });
-
-  it('combined --from-dir + --to-dir + --rewrite-path-refs works end-to-end', async () => {
-    const sourceDir = path.join(tempDir, 'source');
-    const targetDir = path.join(tempDir, 'target');
-    await fs.mkdir(path.join(sourceDir, '.cursor', 'rules'), { recursive: true });
-    await fs.mkdir(targetDir, { recursive: true });
-
-    // Create rules with cross-references
-    await fs.writeFile(
-      path.join(sourceDir, '.cursor/rules/a.mdc'),
-      '---\nalwaysApply: true\n---\n\nReference: .cursor/rules/b.mdc'
-    );
-    await fs.writeFile(
-      path.join(sourceDir, '.cursor/rules/b.mdc'),
-      '---\nalwaysApply: true\n---\n\nRule B content'
-    );
-
-    const result = await engine.convert({
-      source: 'cursor',
-      target: 'claude',
-      root: '/unused',
-      sourceRoot: sourceDir,
-      targetRoot: targetDir,
-      rewritePathRefs: true,
-    });
-
-    expect(result.discovered).toHaveLength(2);
-    expect(result.written).toHaveLength(2);
-
-    // Output should be in targetDir
-    const aContent = await fs.readFile(
-      path.join(targetDir, '.claude/rules/a.md'),
-      'utf-8'
-    );
-    expect(aContent).toContain('.claude/rules/b.md');
-    expect(aContent).not.toContain('.cursor/rules/b.mdc');
-
-    // Source should be untouched (no --delete-source)
-    await expect(
-      fs.access(path.join(sourceDir, '.cursor/rules/a.mdc'))
-    ).resolves.not.toThrow();
-  });
-
-  it('AgentSkillIO rewrites SKILL.md body and reference ride-alongs', async () => {
-    // End-to-end coverage for Behaviors 5–8 of the AgentSkillIO +
-    // --rewrite-path-refs fix:
-    //   - SKILL.md body references to ride-along resources are rewritten.
-    //   - All ride-along files are copied to the target regardless of subtree.
-    //   - Content inside scripts/** and references/** is rewritten; content
-    //     inside assets/** and unknown subtrees (e.g. data/**) is left
-    //     byte-for-byte unchanged.
-    //   - Orphan scanning is scoped to scripts/** and references/** only.
-    const skillDir = path.join(tempDir, '.cursor', 'skills', 'check');
-    await fs.mkdir(path.join(skillDir, 'scripts'), { recursive: true });
-    await fs.mkdir(path.join(skillDir, 'references'), { recursive: true });
-    await fs.mkdir(path.join(skillDir, 'assets'), { recursive: true });
-    await fs.mkdir(path.join(skillDir, 'data'), { recursive: true });
-
-    // SKILL.md references its own ride-along resources.
-    await fs.writeFile(
-      path.join(skillDir, 'SKILL.md'),
-      [
-        '---',
-        'description: Check skill',
-        '---',
-        '',
-        '# Check',
-        '',
-        'Run the helper: .cursor/skills/check/scripts/helper.sh',
-        'See notes: .cursor/skills/check/references/NOTES.md',
-        'Raw data blob: .cursor/skills/check/data/blob.bin',
-        'Image: .cursor/skills/check/assets/logo.png',
-        '',
-      ].join('\n'),
-    );
-    // scripts/helper.sh — content references a sibling script + the SKILL.md
-    const scriptBody = [
-      '#!/bin/sh',
-      '# relies on sibling: .cursor/skills/check/scripts/other.sh',
-      '# back-ref to the skill itself: .cursor/skills/check/SKILL.md',
-      'exit 0',
-      '',
-    ].join('\n');
-    await fs.writeFile(path.join(skillDir, 'scripts', 'helper.sh'), scriptBody);
-    await fs.writeFile(path.join(skillDir, 'scripts', 'other.sh'), '#!/bin/sh\nexit 0\n');
-    // references/NOTES.md — referencing a script (should be rewritten).
-    const notesBody = [
-      '# Notes',
-      '',
-      'The helper lives at .cursor/skills/check/scripts/helper.sh',
-      '',
-    ].join('\n');
-    await fs.writeFile(path.join(skillDir, 'references', 'NOTES.md'), notesBody);
-    // assets/logo.png — must pass through verbatim even if it *looks* like
-    // it has a path ref in it. We use a harmless text payload here so
-    // we can byte-compare; the guarantee is "asset bytes preserved".
-    //
-    // The cursor-path fragment below is INTENTIONALLY unmapped (the file
-    // `.cursor/rules/missing-from-asset.mdc` doesn't exist in the source
-    // tree). If a future regression taught `detectOrphans` to scan assets/,
-    // this unmapped ref would show up in an orphan warning — the final
-    // assertion in this test pins that behaviour closed.
-    const orphanFromAsset = '.cursor/rules/missing-from-asset.mdc';
-    const orphanFromData = '.cursor/rules/missing-from-data.mdc';
-    const assetBytes =
-      `binary-ish payload that happens to mention ${orphanFromAsset}\n`;
-    await fs.writeFile(path.join(skillDir, 'assets', 'logo.png'), assetBytes);
-    // data/** is an unknown subtree — also must pass through verbatim.
-    const dataBody =
-      `raw data referencing ${orphanFromData} should not be touched\n`;
-    await fs.writeFile(path.join(skillDir, 'data', 'blob.bin'), dataBody);
-
-    const result = await engine.convert({
-      source: 'cursor',
-      target: 'claude',
-      root: tempDir,
-      rewritePathRefs: true,
-    });
-
-    // All 5 resource files + SKILL.md should be written.
-    const targetSkillDir = path.join(tempDir, '.claude', 'skills', 'check');
-    for (const rel of [
-      'SKILL.md',
-      'scripts/helper.sh',
-      'scripts/other.sh',
-      'references/NOTES.md',
-      'assets/logo.png',
-      'data/blob.bin',
-    ]) {
-      await expect(
-        fs.access(path.join(targetSkillDir, rel)),
-        `expected ${rel} to be written`,
-      ).resolves.not.toThrow();
-    }
-
-    // SKILL.md body: cursor → claude paths for known (rewritable) refs;
-    // assets/ and data/ refs inside SKILL.md body should ALSO be rewritten
-    // (the body is always in scope), the "scope" limitation only applies
-    // to content *inside* ride-along files.
-    const skillMd = await fs.readFile(path.join(targetSkillDir, 'SKILL.md'), 'utf-8');
-    expect(skillMd).toContain('.claude/skills/check/scripts/helper.sh');
-    expect(skillMd).toContain('.claude/skills/check/references/NOTES.md');
-    expect(skillMd).toContain('.claude/skills/check/data/blob.bin');
-    expect(skillMd).toContain('.claude/skills/check/assets/logo.png');
-    expect(skillMd).not.toContain('.cursor/skills/check/scripts/helper.sh');
-    expect(skillMd).not.toContain('.cursor/skills/check/references/NOTES.md');
-
-    // scripts/helper.sh — rewritten (Behavior 7 allowlist includes scripts/)
-    const helper = await fs.readFile(
-      path.join(targetSkillDir, 'scripts', 'helper.sh'),
-      'utf-8',
-    );
-    expect(helper).toContain('.claude/skills/check/scripts/other.sh');
-    expect(helper).toContain('.claude/skills/check/SKILL.md');
-    expect(helper).not.toContain('.cursor/skills/check/scripts/other.sh');
-    expect(helper).not.toContain('.cursor/skills/check/SKILL.md');
-
-    // references/NOTES.md — rewritten (Behavior 7 allowlist includes references/)
-    const notes = await fs.readFile(
-      path.join(targetSkillDir, 'references', 'NOTES.md'),
-      'utf-8',
-    );
-    expect(notes).toContain('.claude/skills/check/scripts/helper.sh');
-    expect(notes).not.toContain('.cursor/skills/check/scripts/helper.sh');
-
-    // assets/logo.png — verbatim passthrough (NOT in Behavior 7 allowlist).
-    const asset = await fs.readFile(
-      path.join(targetSkillDir, 'assets', 'logo.png'),
-      'utf-8',
-    );
-    expect(asset).toBe(assetBytes);
-    expect(asset).toContain(orphanFromAsset);
-
-    // data/blob.bin — verbatim passthrough (unknown subtree).
-    const data = await fs.readFile(
-      path.join(targetSkillDir, 'data', 'blob.bin'),
-      'utf-8',
-    );
-    expect(data).toBe(dataBody);
-    expect(data).toContain(orphanFromData);
-
-    // Behavior 8: orphan detection should NOT fire just because assets/
-    // or data/ still contain un-rewritten cursor paths. Those subtrees
-    // are deliberately out of scope for path scanning.
-    //
-    // We assert on the *found path* (embedded in the warning message via
-    // `Orphan path reference: '${foundPath}' is not in the conversion set`)
-    // rather than on the subtree name. The asset/data payloads above contain
-    // unmapped cursor paths that WOULD be reported as orphans if their
-    // subtrees were scanned; pinning their absence here is a proper
-    // regression guard. An assertion on `assets/|data/` would be vacuously
-    // true — orphan warnings report the found *path*, not the file that
-    // contained it.
-    const orphanWarnings = result.warnings.filter(
-      (w) => w.code === WarningCode.OrphanPathRef,
-    );
-    const orphanMessages = orphanWarnings.map((w) => w.message).join('\n');
-    expect(
-      orphanMessages,
-      `unmapped ref inside assets/ leaked into an orphan warning: ${orphanMessages}`,
-    ).not.toContain(orphanFromAsset);
-    expect(
-      orphanMessages,
-      `unmapped ref inside data/ leaked into an orphan warning: ${orphanMessages}`,
-    ).not.toContain(orphanFromData);
-  });
-});
-
-describe('Integration Tests - a16n IR Plugin', () => {
-  beforeEach(async () => {
-    await fs.rm(tempDir, { recursive: true, force: true });
-    await fs.mkdir(tempDir, { recursive: true });
-  });
-
-  afterEach(async () => {
-    await fs.rm(tempDir, { recursive: true, force: true });
-  });
-
-  describe('a16n discovery', () => {
-    it('should discover items from .a16n/ directory', async () => {
-      const fixturePath = path.join(fixturesDir, 'a16n-basic');
-      const fromDir = path.join(fixturePath, 'from-a16n');
-
-      // Copy input to temp
-      await copyDir(fromDir, tempDir);
-
-      // Run conversion (discover from a16n, emit to cursor)
-      const result = await engine.convert({
-        source: 'a16n',
-        target: 'cursor',
-        root: tempDir,
-      });
-
-      // Verify items were discovered
-      expect(result.discovered.length).toBeGreaterThan(0);
-    });
-
-    it('should discover all 6 IR types from .a16n/', async () => {
-      const fixturePath = path.join(fixturesDir, 'a16n-basic');
-      const fromDir = path.join(fixturePath, 'from-a16n');
-
-      // Copy input to temp
-      await copyDir(fromDir, tempDir);
-
-      // Run conversion
-      const result = await engine.convert({
-        source: 'a16n',
-        target: 'cursor',
-        root: tempDir,
-      });
-
-      // Verify all 6 types are discovered
-      const types = new Set(result.discovered.map(d => d.type));
-      expect(types.has('global-prompt')).toBe(true);
-      expect(types.has('file-rule')).toBe(true);
-      expect(types.has('simple-agent-skill')).toBe(true);
-      expect(types.has('agent-ignore')).toBe(true);
-      expect(types.has('manual-prompt')).toBe(true);
-      expect(types.has('agent-skill-io')).toBe(true);
-      expect(result.discovered).toHaveLength(6);
-    });
-  });
-
-  describe('a16n-to-cursor conversion', () => {
-    it('should convert a16n IR to Cursor format', async () => {
-      const fixturePath = path.join(fixturesDir, 'a16n-basic');
-      const fromDir = path.join(fixturePath, 'from-a16n');
-
-      // Copy input to temp
-      await copyDir(fromDir, tempDir);
-
-      // Run conversion
-      const result = await engine.convert({
-        source: 'a16n',
-        target: 'cursor',
-        root: tempDir,
-      });
-
-      expect(result.discovered).toHaveLength(6);
-      expect(result.written.length).toBeGreaterThan(0);
-
-      // Verify Cursor rules were created
-      const rulesDir = path.join(tempDir, '.cursor', 'rules');
-      const ruleFiles = await readDirFiles(rulesDir);
-      expect(ruleFiles.size).toBeGreaterThan(0);
-
-      // Check GlobalPrompt was emitted as alwaysApply Cursor rule
-      const ruleContents = Array.from(ruleFiles.values());
-      const hasGlobalPrompt = ruleContents.some(c => c.includes('Always use TypeScript'));
-      expect(hasGlobalPrompt).toBe(true);
-    });
-  });
-
-  describe('a16n-to-claude conversion', () => {
-    it('should convert a16n IR to Claude format', async () => {
-      const fixturePath = path.join(fixturesDir, 'a16n-basic');
-      const fromDir = path.join(fixturePath, 'from-a16n');
-
-      // Copy input to temp
-      await copyDir(fromDir, tempDir);
-
-      // Run conversion
-      const result = await engine.convert({
-        source: 'a16n',
-        target: 'claude',
-        root: tempDir,
-      });
-
-      expect(result.discovered).toHaveLength(6);
-      expect(result.written.length).toBeGreaterThan(0);
-
-      // Verify Claude rules were created
-      const rulesDir = path.join(tempDir, '.claude', 'rules');
-      const ruleFiles = await readDirFiles(rulesDir);
-      expect(ruleFiles.size).toBeGreaterThan(0);
-
-      // Check GlobalPrompt was emitted as Claude rule
-      const ruleContents = Array.from(ruleFiles.values());
-      const hasGlobalPrompt = ruleContents.some(c => c.includes('Always use TypeScript'));
-      expect(hasGlobalPrompt).toBe(true);
-    });
-  });
-
-  describe('round-trip: cursor → a16n → cursor', () => {
-    it('should preserve content through cursor → a16n → cursor round-trip', async () => {
-      // Create a cursor project with a simple rule
-      const cursorRulesDir = path.join(tempDir, '.cursor', 'rules');
-      await fs.mkdir(cursorRulesDir, { recursive: true });
-      await fs.writeFile(
-        path.join(cursorRulesDir, 'coding.mdc'),
-        `---
-alwaysApply: true
----
-
-Always use TypeScript strict mode.
-Follow DRY principles.
-`,
-        'utf-8'
-      );
-
-      // Step 1: Convert Cursor → a16n
-      const toA16n = await engine.convert({
-        source: 'cursor',
-        target: 'a16n',
-        root: tempDir,
-      });
-      expect(toA16n.discovered.length).toBeGreaterThan(0);
-      expect(toA16n.written.length).toBeGreaterThan(0);
-
-      // Verify .a16n/ was created
-      const a16nDir = path.join(tempDir, '.a16n');
-      const a16nStat = await fs.stat(a16nDir);
-      expect(a16nStat.isDirectory()).toBe(true);
-
-      // Step 2: Remove original Cursor rules to prove round-trip works
-      await fs.rm(path.join(tempDir, '.cursor'), { recursive: true, force: true });
-
-      // Step 3: Convert a16n → Cursor
-      const toCursor = await engine.convert({
-        source: 'a16n',
-        target: 'cursor',
-        root: tempDir,
-      });
-      expect(toCursor.discovered.length).toBeGreaterThan(0);
-      expect(toCursor.written.length).toBeGreaterThan(0);
-
-      // Verify Cursor rules were recreated with the original content preserved
-      const restoredRulesDir = path.join(tempDir, '.cursor', 'rules');
-      const restoredFiles = await readDirFiles(restoredRulesDir);
-      expect(restoredFiles.size).toBeGreaterThan(0);
-
-      const restoredContent = Array.from(restoredFiles.values()).join('\n');
-      expect(restoredContent).toContain('Always use TypeScript strict mode');
-      expect(restoredContent).toContain('Follow DRY principles');
-    });
-  });
-
-  describe('round-trip: claude → a16n → claude', () => {
-    it('should preserve content through claude → a16n → claude round-trip', async () => {
-      // Create a Claude project with a rule
-      const claudeRulesDir = path.join(tempDir, '.claude', 'rules');
-      await fs.mkdir(claudeRulesDir, { recursive: true });
-      await fs.writeFile(
-        path.join(claudeRulesDir, 'standards.md'),
-        `---
-trigger: always
----
-
-Write tests for all functions.
-Use meaningful variable names.
-`,
-        'utf-8'
-      );
-
-      // Step 1: Convert Claude → a16n
-      const toA16n = await engine.convert({
-        source: 'claude',
-        target: 'a16n',
-        root: tempDir,
-      });
-      expect(toA16n.discovered.length).toBeGreaterThan(0);
-      expect(toA16n.written.length).toBeGreaterThan(0);
-
-      // Verify .a16n/ was created
-      const a16nDir = path.join(tempDir, '.a16n');
-      const a16nStat = await fs.stat(a16nDir);
-      expect(a16nStat.isDirectory()).toBe(true);
-
-      // Step 2: Remove original Claude rules
-      await fs.rm(path.join(tempDir, '.claude'), { recursive: true, force: true });
-
-      // Step 3: Convert a16n → Claude
-      const toClaude = await engine.convert({
-        source: 'a16n',
-        target: 'claude',
-        root: tempDir,
-      });
-      expect(toClaude.discovered.length).toBeGreaterThan(0);
-      expect(toClaude.written.length).toBeGreaterThan(0);
-
-      // Verify Claude rules were recreated with original content preserved
-      const restoredRulesDir = path.join(tempDir, '.claude', 'rules');
-      const restoredFiles = await readDirFiles(restoredRulesDir);
-      expect(restoredFiles.size).toBeGreaterThan(0);
-
-      const restoredContent = Array.from(restoredFiles.values()).join('\n');
-      expect(restoredContent).toContain('Write tests for all functions');
-      expect(restoredContent).toContain('Use meaningful variable names');
-    });
-  });
-
-  describe('cross-format: cursor → a16n → claude', () => {
-    it('should preserve content through cursor → a16n → claude round-trip', async () => {
-      // Create a Cursor project with a rule
-      const cursorRulesDir = path.join(tempDir, '.cursor', 'rules');
-      await fs.mkdir(cursorRulesDir, { recursive: true });
-      await fs.writeFile(
-        path.join(cursorRulesDir, 'style.mdc'),
-        `---
-alwaysApply: true
----
-
-Use consistent naming conventions.
-Prefer const over let.
-`,
-        'utf-8'
-      );
-
-      // Step 1: Convert Cursor → a16n (IR)
-      const toA16n = await engine.convert({
-        source: 'cursor',
-        target: 'a16n',
-        root: tempDir,
-      });
-      expect(toA16n.discovered.length).toBeGreaterThan(0);
-      expect(toA16n.written.length).toBeGreaterThan(0);
-
-      // Step 2: Remove Cursor source to prove the IR carries the content
-      await fs.rm(path.join(tempDir, '.cursor'), { recursive: true, force: true });
-
-      // Step 3: Convert a16n → Claude
-      const toClaude = await engine.convert({
-        source: 'a16n',
-        target: 'claude',
-        root: tempDir,
-      });
-      expect(toClaude.discovered.length).toBeGreaterThan(0);
-      expect(toClaude.written.length).toBeGreaterThan(0);
-
-      // Verify Claude rules contain the original Cursor content
-      const claudeRulesDir = path.join(tempDir, '.claude', 'rules');
-      const claudeFiles = await readDirFiles(claudeRulesDir);
-      expect(claudeFiles.size).toBeGreaterThan(0);
-
-      const claudeContent = Array.from(claudeFiles.values()).join('\n');
-      expect(claudeContent).toContain('Use consistent naming conventions');
-      expect(claudeContent).toContain('Prefer const over let');
-    });
-  });
-
-  describe('cross-format: claude → a16n → cursor', () => {
-    it('should preserve content through claude → a16n → cursor round-trip', async () => {
-      // Create a Claude project with a rule
-      const claudeRulesDir = path.join(tempDir, '.claude', 'rules');
-      await fs.mkdir(claudeRulesDir, { recursive: true });
-      await fs.writeFile(
-        path.join(claudeRulesDir, 'quality.md'),
-        `---
-trigger: always
----
-
-Always write documentation for public APIs.
-Use descriptive error messages.
-`,
-        'utf-8'
-      );
-
-      // Step 1: Convert Claude → a16n (IR)
-      const toA16n = await engine.convert({
-        source: 'claude',
-        target: 'a16n',
-        root: tempDir,
-      });
-      expect(toA16n.discovered.length).toBeGreaterThan(0);
-      expect(toA16n.written.length).toBeGreaterThan(0);
-
-      // Step 2: Remove Claude source to prove the IR carries the content
-      await fs.rm(path.join(tempDir, '.claude'), { recursive: true, force: true });
-
-      // Step 3: Convert a16n → Cursor
-      const toCursor = await engine.convert({
-        source: 'a16n',
-        target: 'cursor',
-        root: tempDir,
-      });
-      expect(toCursor.discovered.length).toBeGreaterThan(0);
-      expect(toCursor.written.length).toBeGreaterThan(0);
-
-      // Verify Cursor rules contain the original Claude content
-      const cursorRulesDir = path.join(tempDir, '.cursor', 'rules');
-      const cursorFiles = await readDirFiles(cursorRulesDir);
-      expect(cursorFiles.size).toBeGreaterThan(0);
-
-      const cursorContent = Array.from(cursorFiles.values()).join('\n');
-      expect(cursorContent).toContain('Always write documentation for public APIs');
-      expect(cursorContent).toContain('Use descriptive error messages');
-    });
-  });
-
-  describe('version mismatch warning', () => {
-    it('should emit VersionMismatch warning for incompatible IR versions', async () => {
-      // Create a .a16n/ directory with a file that has an incompatible version
-      // v1beta99 is valid but incompatible with current v1beta1 (revision too high)
-      const a16nDir = path.join(tempDir, '.a16n', 'global-prompt');
-      await fs.mkdir(a16nDir, { recursive: true });
-      await fs.writeFile(
-        path.join(a16nDir, 'future-rule.md'),
-        `---
-version: v1beta99
-type: global-prompt
----
-
-This rule is from a future revision.
-`,
-        'utf-8'
-      );
-
-      // Run discovery (a16n → cursor)
-      const result = await engine.convert({
-        source: 'a16n',
-        target: 'cursor',
-        root: tempDir,
-      });
-
-      // Should still discover the item (processed despite mismatch)
-      expect(result.discovered.length).toBeGreaterThan(0);
-
-      // Should have emitted a VersionMismatch warning
-      const versionWarnings = result.warnings.filter(
-        w => w.code === WarningCode.VersionMismatch
-      );
-      expect(versionWarnings).toHaveLength(1);
-      expect(versionWarnings[0].message).toContain('v1beta99');
-      expect(versionWarnings[0].message).toContain('v1beta2');
-    });
-  });
-});
